# src/main/resources/api.yaml
openapi: 3.0.0
servers:
  - description: Local deployment
    url: http://localhost:5000
info:
  description: API to integrate with Google's calendar service
  version: "0.0.1"
  title: Calendar Integration Service
paths:
...
  /events:
    get:
      summary: lists events for given date range (or today)
      security:
        - bearerAuth: []
      operationId: nextEvents
      parameters:
        - in: query
          name: start
          description: inclusive, defaults to today
          schema:
            type: string
            format: date
        - in: query
          name: end
          description: inclusive, defaults to today
          schema:
            type: string
            format: date
        - in: query
          name: timezone
          description: specifies timezone for events, defaults to UTC
          schema:
            type: string
        - in: query
          name: nextToken
          description: token pointing to next page of results
          schema:
            type: string
      responses:
        '200':
          description: A new access token object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NextEvents'
components:
  schemas:
...
    Event:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        internalAttendeesOnly:
          type: boolean
        attendees:
          type: array
          items:
            type: string
            format: email
    NextEvents:
      type: object
      properties:
        nextToken:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
...